requirements.txt
# Web Framework
fastapi>=0.110
uvicorn[standard]>=0.27

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
pgvector>=0.3
alembic>=1.13

# Cache
redis[hiredis]>=5.0

# Configuration
pydantic-settings>=2.1
email-validator>=2.1

# Auth (Phase 2)
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt==4.2.1

# AI / LLM (Phase 3+)
langchain>=0.2
langchain-core>=0.2
langchain-community>=0.2
langchain-openai>=0.1
httpx>=0.27

# Testing
pytest>=8.0
pytest-asyncio>=0.23
pytest-mock>=3.12
pytest-timeout>=2.3
aiosqlite>=0.20
httpx  # also used for TestClient

# Observability (Phase 6)
arize-phoenix[otel]>=4.0
openinference-instrumentation-langchain>=0.1
prometheus-fastapi-instrumentator>=6.1

# Evaluation (Phase 6 CD)
ragas>=0.1.21
datasets>=2.18

# Logging
python-json-logger>=2.0

# Document Processing
pymupdf>=1.24
striprtf>=0.0.26
